do not use namespace std
Doing so may prevent seamless integration between this and other
projects. This happens because one may have a class, in a namespace,
that shares the same name as one of the std classes. When including
the tdigest hpp files, that would create conflicts involving those
classes.

"using namespace xxx" should be fine in .cpp files.